+++ data/media/artifacts/hosting/playground/deployments/6bed1be8-3a6d-479b-bfb6-d4a32109a200/extracted/assets/UserIcon-CwZR14FY.js
@@ -1,2 +1,0 @@
-import{j as o}from"./main-KLk480A7.js";const s=()=>o.jsxs("svg",{xmlns:"http://www.w3.org/2000/svg",width:"16",height:"16",viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round",children:[o.jsx("path",{d:"M20 21a8 8 0 1 0-16 0"}),o.jsx("circle",{cx:"12",cy:"7",r:"4"})]});export{s as U};
-//# sourceMappingURL=UserIcon-CwZR14FY.js.map