+++ data/media/artifacts/hosting/playground/deployments/6bed1be8-3a6d-479b-bfb6-d4a32109a200/extracted/assets/buildAuthUrlWithNext-DGDR15ZY.js
@@ -1,2 +1,0 @@
-const n=o=>!!(o&&o!=="/"),r=(o,e)=>n(e)?`${o}?next=${encodeURIComponent(e)}`:o;export{r as b};
-//# sourceMappingURL=buildAuthUrlWithNext-DGDR15ZY.js.map